requirements.txt
selenium==4.18.1
undetected-chromedriver==3.5.5
selenium-stealth==1.0.6
openai==1.14.0
apscheduler==3.10.4
pyyaml==6.0.1
python-dotenv==1.0.1
sqlalchemy==2.0.28
requests==2.31.0
pytest==8.1.1
pytest-mock==3.12.0
